+++ src/ta_material.h
@@ -2,7 +2,7 @@
 #include "ta_uid.h"
 
 typedef struct ta_material {
-    u32 index;
+    size_t index;
     const char *name;
     const char *shader;
     const char *tex_albedo;
